Refactor detail page
