+++ src/app/modules/project/details/bottom-drawer/bottom-drawer.component.html
@@ -1,0 +1,40 @@
+<div class="bottom">
+  <div class="menu-wrapper">
+    <div class="menu">
+      <div (click)="setActiveTab('description')" [ngClass]="{ active: this.activeTab == 'description' }" class="item">Description</div>
+      <div (click)="setActiveTab('collaborators')" [ngClass]="{ active: this.activeTab == 'collaborators' }" class="item">Collaborators
+      </div>
+    </div>
+    <app-settings-dropdown
+        (projectHighlighted)="isProjectHighlighted = $event"
+        [currentUser]="currentUser"
+        [project]="project"
+    ></app-settings-dropdown>
+  </div>
+
+  <div class="tab-content">
+    <ng-container *ngIf="this.activeTab == 'description'" class="project-description large">
+      <div class="top-buttons">
+        <ng-container *ngIf="displayCallToActionButton">
+          <div class="callToActionButtons">
+            <button (click)="onClickCallToActionButton(cta.value)" *ngFor="let cta of project?.callToActions" class="primary" id="call-to-action-button">
+              {{ cta.optionValue }}
+              <em [classList]="'fas ' + cta.iconName"></em>
+            </button>
+          </div>
+        </ng-container>
+        <ng-container *ngIf="displayEmbedButton">
+          <app-embed-button [projectId]="project.id"></app-embed-button>
+        </ng-container>
+      </div>
+
+
+      <p [innerHTML]="project.description | safeHtml" class="dex-quill-viewer"></p>
+    </ng-container>
+    <ng-container *ngIf="this.activeTab == 'collaborators'">
+      <div class="collaborators">
+        <app-collaborator *ngFor="let collaborator of project.collaborators" [collaborator]="collaborator"></app-collaborator>
+      </div>
+    </ng-container>
+  </div>
+</div>
