+++ src/app/modules/project/details/details.component.scss
@@ -62,6 +62,7 @@
     width: fit-content;
     margin: 0 auto;
     max-width: 100%;
+    border-radius: .3rem;
   }
 }
